example_project/requirements.txt
django-media-fixtures>=0.0.1